Modularize jira/client.rs into submodules for maintainability
Split the 636-line client.rs into focused domain modules:
- client.rs (291 lines): Core HTTP infrastructure and auth
- issues.rs: Issue conversion, listing, and creation
- comments.rs: Comment fetching operations
- goals.rs: Version/goal operations
- fields.rs: Field discovery for JQL queries
